Add checkbox for cc saving
